upgrade: check-diff to display package diff for new version
    This new option has the same functionality as check-only, but if a
    new version is available, it will download but not deploy it, and
    print a package-level diff.  This allows sysadmins to see exactly
    which packages will be changed on upgrade.  If the user does then
    decide to upgrade, it prevents re-downloading the verison by
    deploying the already-downloaded data under the same checksum
    with an incremented deploy serial.

upgrade: Revise check-diff with fixes and eliminate a for loop

This update fixes some small bugs including checking for error and
moving small pieces of code.  It also elimates the need to cycle
through checksums of refs by calling ostree_repo_load_variant_if_exists
and seeing if it succeeds in loading an object.

+++ src/rpmostree-builtin-upgrade.c
@@ -34,16 +34,107 @@
 
 #include "ostree-sysroot-upgrader.h"
 
+/**
+ * SECTION:libostree-sysroot-upgrader
+ * @title: Simple upgrade class
+ * @short_description: Upgrade OSTree systems
+ *
+ * The #OstreeSysrootUpgrader class allows performing simple upgrade
+ * operations.
+ */
+ /*
+ This struct is copy-pasted from ostree-sysroot-upgrader.c
+ because without it, opt_check_diff would not work
+ */
+struct OstreeSysrootUpgrader {
+  GObject parent;
+
+  OstreeSysroot *sysroot;
+  char *osname;
+
+  OstreeDeployment *merge_deployment;
+  GKeyFile *origin;
+  char *origin_remote;
+  char *origin_ref;
+
+  char *new_revision;
+}; 
+
 static gboolean opt_reboot;
 static gboolean opt_allow_downgrade;
 static gboolean opt_check_only;
+static gboolean opt_check_diff;
 
 static GOptionEntry option_entries[] = {
   { "reboot", 'r', 0, G_OPTION_ARG_NONE, &opt_reboot, "Initiate a reboot after an upgrade is prepared", NULL },
   { "allow-downgrade", 0, 0, G_OPTION_ARG_NONE, &opt_allow_downgrade, "Permit deployment of chronologically older trees", NULL },
   { "check-only", 'c', 0, G_OPTION_ARG_NONE, &opt_check_only, "Check for upgrades and print true/false only", NULL },
+  { "check-diff", 0, 0, G_OPTION_ARG_NONE, &opt_check_diff, "Check for upgrades and print package diff only", NULL },
   { NULL }
 };
+
+/**
+ * given a repo and an origin_description,
+ * this uses libsoup to access the remote
+ * repo and read the checksum of the 
+ * commit version
+ */
+static gboolean
+get_checksum_from_repo (OstreeRepo    *repo,
+                        char          *origin_description,
+                        char         **out_checksum,
+                        GCancellable  *cancellable,
+                        GError       **error)
+{
+  gboolean ret = FALSE;
+
+  gs_free char *checksum_from_file = NULL;
+  GKeyFile *config = NULL; // config file, holds URL
+  GDataInputStream *data_in = NULL;
+  GInputStream *gis = NULL;
+  char *inbetween = NULL; // necessary addition to URI
+  gs_free char *key = NULL; // key of url in config
+  gsize length; // length of checksum in file
+  gs_free char *ref = NULL; // location of this rev
+  gs_free char *remote = NULL;
+  SoupRequest *req = NULL;
+  SoupSession *session = soup_session_new();
+  gs_free char *uri = NULL; // full uri location of this ref
+  gs_free char *url = NULL; // url from config file
+
+  if (!ostree_parse_refspec (origin_description, &remote, &ref, error))
+        goto out;
+
+  config = ostree_repo_copy_config (repo);
+  key = g_strdup_printf ("remote \"%s\"", remote);
+  url = g_key_file_get_string (config, key, "url", error);
+  if (url == NULL)
+    goto out;
+
+  inbetween = "/refs/heads/";
+  uri = g_strconcat (url, inbetween, ref, NULL);
+
+  req = soup_session_request (session, uri, error);
+  if (!req)
+    goto out;
+
+  gis = soup_request_send (req, cancellable, error);
+  if (!gis)
+    goto out;
+  
+  data_in = g_data_input_stream_new (gis);
+  if (!data_in)
+    goto out;
+
+  checksum_from_file =  g_data_input_stream_read_line  (data_in, &length, cancellable, error);
+  if (!checksum_from_file)
+    goto out;
+
+  ret = TRUE;
+  gs_transfer_out_value (out_checksum, &checksum_from_file);
+ out:
+  return ret;
+}
 
 gboolean
 rpmostree_builtin_upgrade (int             argc,
@@ -59,25 +150,14 @@ rpmostree_builtin_upgrade (int             argc,
   GSConsole *console = NULL;
   gboolean changed;
   OstreeSysrootUpgraderPullFlags upgraderpullflags = 0;
-  gs_free char *origin_description = NULL;
-  guint i;
 
-  GKeyFile *config = NULL; // config file, holds URL
-  gs_free char *url = NULL; // url from config file
-  gs_free char *remote = NULL;
-  gs_free char *ref = NULL; // location of this rev
-  char *inbetween = NULL; // necessary addition to URI
-  gs_unref_object OstreeRepo *repo = NULL; 
-  gs_free char *key = NULL; // key of url in config
-  gs_free char *uri = NULL; // full uri location of this ref
-  gsize length; // length of checksum in file
-  gs_free char *data_from_file = NULL; // the checksum libsoup reads
-  GInputStream *gis = NULL;
-  GDataInputStream *data_in = NULL;
-  SoupSession *session = soup_session_new();
-  SoupRequest *req = NULL;
+  gboolean already_deployed = FALSE;
+  gboolean already_downloaded = FALSE;
   GPtrArray *deployments = NULL;  // list of all depoyments
   gboolean found_csum_match = FALSE;
+  guint i, j;
+  gs_free char *origin_description = NULL;
+  gs_unref_object OstreeRepo *repo = NULL; 
 
   g_option_context_add_main_entries (context, option_entries, NULL);
 
@@ -99,34 +179,18 @@ rpmostree_builtin_upgrade (int             argc,
   if (!ostree_sysroot_get_repo (sysroot, &repo, cancellable, error))
         goto out;
 
-  if (!ostree_parse_refspec (origin_description, &remote, &ref, error))
-        goto out;
+  deployments = ostree_sysroot_get_deployments (sysroot);
+
+  if (opt_check_diff)
+    opt_check_only = TRUE;
 
   if (opt_check_only)
     {
-      config = ostree_repo_copy_config (repo);
-      key = g_strdup_printf ("remote \"%s\"", remote);
-      url = g_key_file_get_string (config, key, "url", error);
-      if (url == NULL)
+      char *data_from_file = NULL; // the checksum libsoup reads
+
+      if (!get_checksum_from_repo (repo, origin_description, &data_from_file, cancellable, error))
         goto out;
 
-      inbetween = "/refs/heads/";
-      uri = g_strdup_printf ("%s%s%s", url, inbetween, ref);
-
-      req = soup_session_request (session, uri, error);
-      if (!req)
-        goto out;
-
-      gis = soup_request_send (req, cancellable, error);
-      if (!gis)
-        goto out;
-      
-      data_in = g_data_input_stream_new (gis);
-      if (!data_in)
-        goto out;
-
-      data_from_file =  g_data_input_stream_read_line  (data_in, &length, cancellable, error);
-      deployments = ostree_sysroot_get_deployments (sysroot);
       for (i=0; i < deployments->len; i++)
         {
           const char *csum = ostree_deployment_get_csum (deployments->pdata[i]);
@@ -137,8 +201,11 @@ rpmostree_builtin_upgrade (int             argc,
       g_print ("Update available: %s\n", found_csum_match ? "false" : "true");
     }
 
-  else
+  if ((opt_check_diff && !found_csum_match) || !opt_check_only)
     {
+      char *checksum_from_repo;
+      gs_unref_variant GVariant *version = NULL;
+
       console = gs_console_get ();
       if (console)
         {
@@ -149,9 +216,56 @@ rpmostree_builtin_upgrade (int             argc,
       if (opt_allow_downgrade)  
         upgraderpullflags |= OSTREE_SYSROOT_UPGRADER_PULL_FLAGS_ALLOW_OLDER;
 
-      if (!ostree_sysroot_upgrader_pull (upgrader, 0, 0, progress, &changed,
-                                         cancellable, error))
+      if (!get_checksum_from_repo (repo, origin_description, &checksum_from_repo, cancellable, error))
         goto out;
+
+      if (!ostree_repo_load_variant_if_exists (repo,
+                                               OSTREE_OBJECT_TYPE_COMMIT,
+                                               checksum_from_repo,
+                                               &version,
+                                               error))
+            goto out;
+      
+      /* If ostree_repo_load_variant_if_exists returned version as
+      anything other than null, then the commit exists and we've already
+      downloaded this version */
+      if (version)
+        {
+          already_downloaded = TRUE;
+          for (j=0; j < deployments->len; j++)
+              {
+                /* this checks if the repo checksum matches any of the DEPLOYMENTS
+                (instead of just ref checksums) */
+                const char *csum = ostree_deployment_get_csum (deployments->pdata[j]);
+                if (g_strcmp0 (csum, checksum_from_repo) == 0)
+                  already_deployed = TRUE;
+              }
+        }
+      
+      /* only pull if we haven't already downloaded
+         OR if we already deployed, which means either:
+         a) current deployment is most recent version, "changed" will be false, or
+         b)  we rolled back and current dpeloyment is not most recent - following
+         normal functionality of upgrade it will pull and deploy the upgrade with
+         an incremented deploy serial 
+         NOTE - in situation b running --check-only will indicate "update available: false" */
+      if (!already_downloaded || already_deployed)
+        {
+          if (!ostree_sysroot_upgrader_pull (upgrader, 0, 0, progress, &changed,
+                                           cancellable, error))
+            goto out;
+        }
+      /* update the upgrader with the checksum of the already-downloaded version
+      this would have been down in upgrader_pull but we skipped that 
+      this way, the upgrader knows what to deploy in deploy_tree */
+      else
+        {
+          char *checksum_of_download = NULL;
+          if (!get_checksum_from_repo (repo, origin_description, &checksum_of_download, cancellable, error))
+            goto out;
+          upgrader->new_revision = checksum_of_download;
+          changed = TRUE;
+        }
 
       if (console)
         {
@@ -169,21 +283,48 @@ rpmostree_builtin_upgrade (int             argc,
         }
       else
         {
-          if (!ostree_sysroot_upgrader_deploy (upgrader, cancellable, error))
-            goto out;
+          if (!opt_check_diff)
+            {
+              if (!ostree_sysroot_upgrader_deploy (upgrader, cancellable, error))
+                    goto out;
+            
+              if (opt_reboot)
+                gs_subprocess_simple_run_sync (NULL, GS_SUBPROCESS_STREAM_DISPOSITION_INHERIT,
+                                               cancellable, error,
+                                               "systemctl", "reboot", NULL);
+              else
+                {
+#ifdef HAVE_PATCHED_HAWKEY_AND_LIBSOLV
+                  if (!rpmostree_print_treepkg_diff (sysroot, cancellable, error))
+                    goto out;
+#endif
 
-          if (opt_reboot)
-            gs_subprocess_simple_run_sync (NULL, GS_SUBPROCESS_STREAM_DISPOSITION_INHERIT,
-                                           cancellable, error,
-                                           "systemctl", "reboot", NULL);
+                  g_print ("Updates prepared for next boot; run \"systemctl reboot\" to start a reboot\n");
+                }
+            }
+          /* check-diff option that does NOT write deployment
+          instead just prints package diff */
           else
             {
-    #ifdef HAVE_PATCHED_HAWKEY_AND_LIBSOLV
-              if (!rpmostree_print_treepkg_diff (sysroot, cancellable, error))
-                goto out;
-    #endif
+              OstreeDeployment *new_deployment = NULL;
+              OstreeDeployment *current_deployment = ostree_sysroot_get_booted_deployment (sysroot);
+              gs_unref_object GFile *current_root = NULL;
+              gs_unref_object GFile *new_root = NULL;
 
-              g_print ("Updates prepared for next boot; run \"systemctl reboot\" to start a reboot\n");
+              if (!ostree_sysroot_deploy_tree (upgrader->sysroot, upgrader->osname,
+                                               upgrader->new_revision,
+                                               upgrader->origin,
+                                               upgrader->merge_deployment,
+                                               NULL,
+                                               &new_deployment,
+                                               cancellable, error))
+                goto out;
+
+              current_root = ostree_sysroot_get_deployment_directory (sysroot, current_deployment);
+              new_root = ostree_sysroot_get_deployment_directory (sysroot, new_deployment);
+
+              if (!print_rpmdb_diff (current_root, new_root, cancellable, error))
+                goto out;
             }
         }
     }
